// frontend/src/types/keyType.ts
type keyType =
	| "rating"
	| "league games"
	| "cup games"
	| "ucl games"
	| "total games"
	| "league goals"
	| "cup goals"
	| "ucl goals"
	| "total goals"
	| "league assists"
	| "cup assists"
	| "ucl assists"
	| "total assists"
	| "league contributions"
	| "cup contributions"
	| "ucl contributions"
	| "total contributions"
	| "goals per game"
	| "assists per game"
	| "contributions per game"
	| "clean sheets"
	| "clean sheets per game"
	| "hattricks"
	| "yellow cards"
	| "red cards"
	| "age"
	| "status";

export default keyType;