// frontend/src/types/valueType.ts
type valueType =
	| "name"
	| "photo"
	| "age"
	| "nationality"
	| "position"
	| "updatePage"
	| "leagueGoals"
	| "cupGoals"
	| "uclGoals"
	| "totalGoals"
	| "leagueGames"
	| "cupGames"
	| "uclGames"
	| "totalGames"
	| "leagueAssists"
	| "cupAssists"
	| "uclAssists"
	| "totalAssists"
	| "leagueContributions"
	| "cupContributions"
	| "uclContributions"
	| "totalContributions"
	| "goalsPerGame"
	| "assistsPerGame"
	| "contributionsPerGame"
	| "cleanSheets"
	| "cleanSheetsPerGame"
	| "hattricks"
	| "rating"
	| "yellowCards"
	| "redCards";

export default valueType;